{{cookiecutter.project_slug}}/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[project]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = "{{cookiecutter.project_short_description}}"
authors = [
    { name = "{{cookiecutter.full_name}}", email = "{{cookiecutter.email}}" },
]
requires-python = ">=3.10"

dependencies = [
    "fastapi[all]>=0.115,<0.116",
    "pydantic>=2.10,<2.11",
    "pydantic-settings>=2.6,<2.7",
    "websockets>=14.0,<15.0",
    "python-dotenv>=1.0,<1.1",
    # deployment libraries
    "uvicorn[standard]>=0.32,<0.33",
    "gunicorn>=23.0,<24.0",
    # others
    "requests>=2.32.3"



]
classifiers = [
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.13",
]


[dependency-groups]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[tool.ruff]
target-version = "py313"
exclude = ["alembic"]

